+++ src/main/java/app/bpartners/api/service/areapicture/AreaPictureAnnotationService.java
@@ -1,24 +1,37 @@
 package app.bpartners.api.service.areapicture;
 
-import app.bpartners.api.endpoint.event.EventProducer;
-import app.bpartners.api.endpoint.event.model.ExportAreaPictureAnnotationRequested;
+import static app.bpartners.api.endpoint.rest.model.FileType.ATTACHMENT;
+import static java.util.UUID.randomUUID;
+
 import app.bpartners.api.endpoint.rest.model.ExportAreaPictureAnnotation;
+import app.bpartners.api.endpoint.rest.model.PreSignedURL;
+import app.bpartners.api.file.FileWriter;
 import app.bpartners.api.model.AreaPictureAnnotation;
 import app.bpartners.api.model.BoundedPageSize;
 import app.bpartners.api.model.PageFromOne;
 import app.bpartners.api.model.exception.NotFoundException;
 import app.bpartners.api.repository.AreaPictureAnnotationRepository;
+import app.bpartners.api.service.annotation.ExportAreaPictureAnnotationPDFProcessor;
+import app.bpartners.api.service.aws.S3Service;
+import java.time.Instant;
 import java.util.List;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+@Slf4j
 @Service
 @AllArgsConstructor
 public class AreaPictureAnnotationService {
+  private static final String PDF_EXTENSION = ".pdf";
+  private static final int ONE_HOUR_IN_SECONDS = 3600;
+
+  private final FileWriter fileWriter;
+  private final S3Service s3Service;
   private final AreaPictureAnnotationRepository repository;
-  private final EventProducer<ExportAreaPictureAnnotationRequested> eventProducer;
+  private final ExportAreaPictureAnnotationPDFProcessor exportAreaPictureAnnotationPDFProcessor;
 
   public AreaPictureAnnotation save(AreaPictureAnnotation areaPictureAnnotation) {
     return repository.save(areaPictureAnnotation);
@@ -70,16 +83,25 @@ public class AreaPictureAnnotationService {
             Sort.by(Sort.Order.desc("creationDatetime"))));
   }
 
-  public ExportAreaPictureAnnotation exportAreaPictureAnnotationToPdf(
+  public PreSignedURL exportAreaPictureAnnotationToPdf(
       String userId, ExportAreaPictureAnnotation annotation) {
-    var request =
-        ExportAreaPictureAnnotationRequested.builder()
-            .annotation(annotation)
-            .userId(userId)
-            .build();
+    try {
+      var generatedPDF = exportAreaPictureAnnotationPDFProcessor.process(annotation);
+      log.info("Pdf generated");
 
-    eventProducer.accept(List.of(request));
+      var fileToUpload = fileWriter.apply(generatedPDF, null);
+      var fileId = "Rapport_d_analyse_" + randomUUID() + PDF_EXTENSION;
 
-    return annotation;
+      s3Service.uploadFile(ATTACHMENT, fileId, userId, fileToUpload);
+      var fileUrl = s3Service.presignURL(ATTACHMENT, fileId, userId, (long) ONE_HOUR_IN_SECONDS);
+
+      return new PreSignedURL()
+          .value(fileUrl)
+          .expirationDelay(ONE_HOUR_IN_SECONDS)
+          .updatedAt(Instant.now());
+    } catch (Exception e) {
+      log.error(e.getMessage());
+      throw new RuntimeException(e.getMessage());
+    }
   }
 }